frontend/requirements.txt
streamlit
requests
python-dotenv
markdown
bleach
